JENKINS: WINSTONE: [JIRA 12854] Add support for large files (> 2 GB) to OutputStream
Code Review Feedback: Andre Gross' proposed changes look straight forward to me in addressing the issue described in JENKINS-12854: Large Artifacts can not be downloaded over WebInterface at https://issues.jenkins-ci.org/browse/JENKINS-12854

+++ src/java/winstone/accesslog/SimpleAccessLogger.java
@@ -81,7 +81,7 @@ public class SimpleAccessLogger implements AccessLogger {
         String uriLine = request.getMethod() + " " + originalURL + " " + request.getProtocol();
         int status = response.getErrorStatusCode() == null ? response.getStatus() 
                 : response.getErrorStatusCode();
-        int size = response.getWinstoneOutputStream().getBytesCommitted();
+        long size = response.getWinstoneOutputStream().getBytesCommitted();
         String date;
         synchronized (DF) {
             date = DF.format(new Date());
